remove the facts_d_directory requires
otherwise it is not possible to use the manage_facts_d_dir switch.

When manage_facts_d_dir is set to false the File[facts_d_directory]
will not be set and the puppet run fails with

"Error: Could not retrieve catalog from remote server: Error 400 on
SERVER: Invalid relationship: File[facts_file] { require =>
File[facts_d_directory] }, because File[facts_d_directory] doesn't seem
to be in the catalog"

Therefore it is necessary to remove this dependency.

The removal of the facts_d_directory requires can cause "No such file or
directory" error. But if you set the manage_facts_d_dir to false it
makes sense to manage the directory somewhere else... Therefore this
error shouldn't occur (if you know what u r doing)